made correction

+++ src/com/company/Circle.java
@@ -4,11 +4,11 @@ public class Circle {
     public static final double PI = 3.14;
 
 
-    public static void area(int radius) {
+    public static void area(double radius) {
         System.out.println("Area of Circle : " + (PI * (radius * radius)));
     }
 
-    public static void cirumference(int radius) {
+    public static void cirumference(double radius) {
         System.out.println("Circumference of Circle : " + (PI * 2 * radius));
     }
